requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cloudinary==1.39.0
Django==5.0.3
django-appconf==1.0.6
django-ckeditor==6.7.1
django-compressor==4.4
django-environ==0.11.2
django-filter==24.1
django-js-asset==2.2.0
django-taggit==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.5.2
packaging==24.0
pillow==10.2.0
pytz==2024.1
PyYAML==6.0.1
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
